LINT in progress

+++ Source/Core/L0Portability/OperatingSystem/Generic/StringHelperExtras_Generic.cpp
@@ -31,7 +31,6 @@
 
 #include "../../StringHelper.h"
 
-
 /*---------------------------------------------------------------------------*/
 /*                           Static definitions                              */
 /*---------------------------------------------------------------------------*/
@@ -43,8 +42,8 @@
 namespace StringHelper {
 
 bool Concatenate(const char8* const string1,
-                                 const char8* const string2,
-                                 char8* const result) {
+                 const char8* const string2,
+                 char8* const result) {
 
     bool ret = false;
 
@@ -82,9 +81,9 @@ bool Concatenate(const char8* const string1,
 }
 
 bool ConcatenateN(const char8* const string1,
-                                  const char8* const string2,
-                                  char8* const result,
-                                  const uint32 size) {
+                  const char8* const string2,
+                  char8* const result,
+                  const uint32 size) {
 
     bool ret = false;
     uint32 sizeCopy = size;
@@ -124,10 +123,9 @@ bool ConcatenateN(const char8* const string1,
     return ret;
 }
 
-
 const char8* TokenizeByChars(const char8* const string,
-                                             const char8* const delimiter,
-                                             char8* const result) {
+                             const char8* const delimiter,
+                             char8* const result) {
 
     const char8 *ret = static_cast<const char8*>(NULL);
 
@@ -171,12 +169,12 @@ const char8* TokenizeByChars(const char8* const string,
 }
 
 const char8* TokenizeByString(const char8* const string,
-                                              const char8* const terminator,
-                                              char8* const result) {
+                              const char8* const terminator,
+                              char8* const result) {
 
     const char8 *ret = static_cast<const char8*>(NULL);
-    int32 size1 = Length(string);
-    int32 size2 = Length(terminator);
+    int32 size1 = static_cast<int32>(Length(string));
+    int32 size2 = static_cast<int32>(Length(terminator));
 
     if ((size1 >= 0) && (size2 >= 0) && (result != NULL)) {
 
@@ -207,9 +205,9 @@ const char8* TokenizeByString(const char8* const string,
 }
 
 bool Substr(const uint32 begin,
-                            const uint32 end,
-                            const char8* const string,
-                            char8* const result) {
+            const uint32 end,
+            const char8* const string,
+            char8* const result) {
 
     bool ret = true;
     if ((string == NULL) || (result == NULL) || (end < begin)) {
